Reworked environment files to use secrets
From now on the environment files are added to the .gitignore file. They aare compiled on-the-fly with a .env file in local machines, and a GitHub secret when deploying
