+++ .github/workflows/firebase-hosting-merge.yml
@@ -12,9 +12,12 @@ jobs:
     steps:
       - uses: actions/checkout@v2
       - run: yarn install --frozen-lockfile && yarn run build
+        env:
+          IMDB_API_KEY_V3: '${{ secrets.IMDB_API_KEY_V3 }}'
+          IMDB_API_KEY_V4: '${{ secrets.IMDB_API_KEY_V4 }}'
       - uses: FirebaseExtended/action-hosting-deploy@v0
         with:
           repoToken: '${{ secrets.GITHUB_TOKEN }}'
           firebaseServiceAccount: '${{ secrets.FIREBASE_SERVICE_ACCOUNT_DOU_DOUM }}'
-          channelId: live
-          projectId: dou-doum
+          channelId: 'live'
+          projectId: 'dou-doum'
