+++ .github/workflows/firebase-hosting-pull-request.yml
@@ -10,8 +10,11 @@ jobs:
     steps:
       - uses: actions/checkout@v2
       - run: yarn install --frozen-lockfile && yarn run build
+        env:
+          IMDB_API_KEY_V3: '${{ secrets.IMDB_API_KEY_V3 }}'
+          IMDB_API_KEY_V4: '${{ secrets.IMDB_API_KEY_V4 }}'
       - uses: FirebaseExtended/action-hosting-deploy@v0
         with:
           repoToken: '${{ secrets.GITHUB_TOKEN }}'
           firebaseServiceAccount: '${{ secrets.FIREBASE_SERVICE_ACCOUNT_DOU_DOUM }}'
-          projectId: dou-doum
+          projectId: 'dou-doum'
