+++ package.json
@@ -3,10 +3,13 @@
   "version": "0.0.0",
   "scripts": {
     "ng": "ng",
+    "prestart": "yarn-or-npm run config --environment=dev",
     "start": "ng serve",
+    "prebuild": "yarn-or-npm run config --environment=prod",
     "build": "ng build",
     "watch": "ng build --watch --configuration development",
-    "test": "ng test"
+    "test": "ng test",
+    "config": "ts-node ./scripts/setenv.ts"
   },
   "private": true,
   "dependencies": {
@@ -37,12 +40,16 @@
     "@angular/compiler-cli": "^14.0.3",
     "@types/jasmine": "~3.10.0",
     "@types/node": "^12.11.1",
+    "dotenv": "^16.0.1",
     "jasmine-core": "~4.0.0",
     "karma": "~6.3.0",
     "karma-chrome-launcher": "~3.1.0",
     "karma-coverage": "~2.1.0",
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "~1.7.0",
-    "typescript": "~4.6.2"
+    "ts-node": "^10.8.1",
+    "typescript": "~4.6.2",
+    "yargs": "^17.5.1",
+    "yarn-or-npm": "^3.0.1"
   }
 }
