+++ src/app/app.component.html
@@ -1,0 +1,1 @@
+<p>{{ env }}</p>
