--- src/environments/environment.prod.ts
@@ -1,12 +1,0 @@
-export const environment = {
-  firebase: {
-    projectId: 'dou-doum',
-    appId: '1:68465041503:web:e3f0c54dde23f9416d3bea',
-    storageBucket: 'dou-doum.appspot.com',
-    apiKey: 'AIzaSyAx2wBByuARIzlI-nGfeMFSIj2a5fgn6NE',
-    authDomain: 'dou-doum.firebaseapp.com',
-    messagingSenderId: '68465041503',
-    measurementId: 'G-M6LLYJJS52',
-  },
-  production: true
-};
