--- src/environments/environment.ts
@@ -1,25 +1,0 @@
-// This file can be replaced during build by using the `fileReplacements` array.
-// `ng build` replaces `environment.ts` with `environment.prod.ts`.
-// The list of file replacements can be found in `angular.json`.
-
-export const environment = {
-  firebase: {
-    projectId: 'dou-doum',
-    appId: '1:68465041503:web:e3f0c54dde23f9416d3bea',
-    storageBucket: 'dou-doum.appspot.com',
-    apiKey: 'AIzaSyAx2wBByuARIzlI-nGfeMFSIj2a5fgn6NE',
-    authDomain: 'dou-doum.firebaseapp.com',
-    messagingSenderId: '68465041503',
-    measurementId: 'G-M6LLYJJS52',
-  },
-  production: false
-};
-
-/*
- * For easier debugging in development mode, you can import the following file
- * to ignore zone related error stack frames such as `zone.run`, `zoneDelegate.invokeTask`.
- *
- * This import should be commented out in production mode because it will have a negative impact
- * on performance if an error is thrown.
- */
-// import 'zone.js/plugins/zone-error';  // Included with Angular CLI.
